E3_MettreDispositionIA/poc_ia/requirements.txt
streamlit==1.32.0
supabase==2.3.0
sentence-transformers==2.5.1
scikit-learn==1.4.0
numpy==1.26.4
Pillow==10.2.0
azure-cognitiveservices-speech==1.34.0 